removed old-assets
