+++ index.js
@@ -1,116 +1,3 @@
-// PARTICLES BACKGROUND
-particlesJS("bg", {
-    particles: {
-      number: {
-        value: 650,
-        density: {
-          enable: true,
-          value_area: 789.1476416322727,
-        },
-      },
-      color: {
-        value: "#ffffff",
-      },
-      shape: {
-        type: "circle",
-        stroke: {
-          width: 0,
-          color: "#000000",
-        },
-        polygon: {
-          nb_sides: 5,
-        },
-        image: {
-          src: "img/github.svg",
-          width: 100,
-          height: 100,
-        },
-      },
-      opacity: {
-        value: 0.48927153781200905,
-        random: false,
-        anim: {
-          enable: true,
-          speed: 1.5,
-          opacity_min: 0,
-          sync: false,
-        },
-      },
-      size: {
-        value: 3,
-        random: true,
-        anim: {
-          enable: true,
-          speed: 2,
-          size_min: 0,
-          sync: false,
-        },
-      },
-      line_linked: {
-        enable: false,
-        distance: 150,
-        color: "#ffffff",
-        opacity: 0.4,
-        width: 1,
-      },
-      move: {
-        enable: true,
-        speed: 0.2,
-        direction: "none",
-        random: true,
-        straight: false,
-        out_mode: "out",
-        bounce: false,
-        attract: {
-          enable: false,
-          rotateX: 600,
-          rotateY: 1200,
-        },
-      },
-    },
-    interactivity: {
-      detect_on: "canvas",
-      events: {
-        onhover: {
-          enable: true,
-          mode: "bubble",
-        },
-        onclick: {
-          enable: true,
-          mode: "push",
-        },
-        resize: true,
-      },
-      modes: {
-        grab: {
-          distance: 400,
-          line_linked: {
-            opacity: 1,
-          },
-        },
-        bubble: {
-          distance: 83.91608391608392,
-          size: 1,
-          duration: 3,
-          opacity: 1,
-          speed: 3,
-        },
-        repulse: {
-          distance: 200,
-          duration: 0.4,
-        },
-        push: {
-          particles_nb: 4,
-        },
-        remove: {
-          particles_nb: 2,
-        },
-      },
-    },
-    retina_detect: true,
-  });
-
-  
   // NASA API CALL 1 : APoD
   
   fetch(
@@ -206,45 +93,3 @@ particlesJS("bg", {
   handleMediaQueryChange(mediaQuery);
   mediaQuery.addListener(handleMediaQueryChange);
   
-  //sidebar toggle script
-  function sidebarToggle() {
-  let menu = document.querySelector(".sidebar-menu");
-  let icon = document.querySelector(".sidebar-icon");
-  
-  function toggleMenu() {
-    menu.classList.toggle("active");
-  }
-  
-  
-  document.addEventListener("DOMContentLoaded", function() {
-    let menu = document.querySelector(".sidebar-menu");
-    let icon = document.querySelector(".sidebar-icon");
-  
-    function toggleMenu() {
-      menu.classList.toggle("active");
-    }
-    icon.addEventListener("click", toggleMenu);
-  
-    let sidebarItems = document.querySelectorAll(".sidebar-menu .rem");
-  
-    sidebarItems.forEach(function(item) {
-      item.addEventListener("click", function() {
-        menu.classList.remove("active");
-      });
-    });
-  });
-}
-  
-  /* scroll to remove the sidebar */
-  var prevScrollpos = window.pageYOffset;
-  window.onscroll = function() {
-    var currentScrollPos = window.pageYOffset;
-    if (prevScrollpos < currentScrollPos) {
-      document.querySelector(".sidebar-menu").classList.remove("active");
-    } else {
-      document.querySelector(".sidebar-menu").classList.remove("active");
-    }
-    prevScrollpos = currentScrollPos;
-  }
-
-  export default sidebarToggle;
