--- old-assets/app.js
@@ -1,220 +1,0 @@
-// PARTICLES BACKGROUND
-
-particlesJS("bg", {
-  particles: {
-    number: {
-      value: 650,
-      density: {
-        enable: true,
-        value_area: 789.1476416322727,
-      },
-    },
-    color: {
-      value: "#ffffff",
-    },
-    shape: {
-      type: "circle",
-      stroke: {
-        width: 0,
-        color: "#000000",
-      },
-      polygon: {
-        nb_sides: 5,
-      },
-      image: {
-        src: "img/github.svg",
-        width: 100,
-        height: 100,
-      },
-    },
-    opacity: {
-      value: 0.48927153781200905,
-      random: false,
-      anim: {
-        enable: true,
-        speed: 1.5,
-        opacity_min: 0,
-        sync: false,
-      },
-    },
-    size: {
-      value: 3,
-      random: true,
-      anim: {
-        enable: true,
-        speed: 2,
-        size_min: 0,
-        sync: false,
-      },
-    },
-    line_linked: {
-      enable: false,
-      distance: 150,
-      color: "#ffffff",
-      opacity: 0.4,
-      width: 1,
-    },
-    move: {
-      enable: true,
-      speed: 0.2,
-      direction: "none",
-      random: true,
-      straight: false,
-      out_mode: "out",
-      bounce: false,
-      attract: {
-        enable: false,
-        rotateX: 600,
-        rotateY: 1200,
-      },
-    },
-  },
-  interactivity: {
-    detect_on: "canvas",
-    events: {
-      onhover: {
-        enable: true,
-        mode: "bubble",
-      },
-      onclick: {
-        enable: true,
-        mode: "push",
-      },
-      resize: true,
-    },
-    modes: {
-      grab: {
-        distance: 400,
-        line_linked: {
-          opacity: 1,
-        },
-      },
-      bubble: {
-        distance: 83.91608391608392,
-        size: 1,
-        duration: 3,
-        opacity: 1,
-        speed: 3,
-      },
-      repulse: {
-        distance: 200,
-        duration: 0.4,
-      },
-      push: {
-        particles_nb: 4,
-      },
-      remove: {
-        particles_nb: 2,
-      },
-    },
-  },
-  retina_detect: true,
-});
-
-// NASA API CALL 1 : APoD
-
-fetch(
-  "https://api.nasa.gov/planetary/apod?api_key=eISHzmGbQk7EMWv9NTlVtvzZVmeLKPUoL3uKMSPJ"
-)
-  .then((response) => {
-    return response.json();
-  })
-  .then((data) => {
-    console.log(data);
-
-    displayData(
-      data.url,
-      data.explanation,
-      data.title,
-      data.date,
-      data.copyright
-    );
-  })
-  .catch((error) => console.error(error));
-
-//function to display data in apod section
-function displayData(image, info, title, date, copyright) {
-  document.getElementById("apod_img").src = image;
-  document.getElementById("apod_info").textContent = info;
-  document.getElementById("title").textContent = title;
-  document.getElementById("date").textContent = date;
-  document.getElementById("copyright").textContent = copyright;
-}
-
-// NASA API CALL 2 : Mars Rover Photos
-
-
-function updateDom(img_src, date, roverName, camera, launch, land, status) {
-  document.querySelector("#roverImg").src = img_src;
-  document.querySelector(".date").textContent = date;
-  document.querySelector(".roverName").textContent = roverName;
-  document.querySelector(".camera").textContent = camera;
-  document.querySelector(".land").textContent = land;
-  document.querySelector(".launch").textContent = launch;
-  document.querySelector(".status").textContent = status;
-}
-
-// API CALL 3 : NASA Photo Video Gallery
-fetch("https://images-api.nasa.gov/search?q=apollo")
-  .then((response) => {
-    return response.json();
-  })
-  .then((data) => {
-    console.log(data);
-  })
-  .catch((error) => console.error(error));
-
-// changed the function and applied media query for apod_img
-function removeClasses() {
-  var element = document.getElementById("class_removal");
-  element.classList.remove("rounded-3", "border");
-}
-
-var mediaQuery = window.matchMedia("(max-width: 765px)");
-function handleMediaQueryChange(mediaQuery) {
-  if (mediaQuery.matches) {
-    removeClasses();
-  }
-}
-
-handleMediaQueryChange(mediaQuery);
-mediaQuery.addListener(handleMediaQueryChange);
-
-//sidebar toggle script
-let menu = document.querySelector(".sidebar-menu");
-let icon = document.querySelector(".sidebar-icon");
-
-function toggleMenu() {
-  menu.classList.toggle("active");
-}
-
-
-document.addEventListener("DOMContentLoaded", function() {
-  let menu = document.querySelector(".sidebar-menu");
-  let icon = document.querySelector(".sidebar-icon");
-
-  function toggleMenu() {
-    menu.classList.toggle("active");
-  }
-  icon.addEventListener("click", toggleMenu);
-
-  let sidebarItems = document.querySelectorAll(".sidebar-menu .rem");
-
-  sidebarItems.forEach(function(item) {
-    item.addEventListener("click", function() {
-      menu.classList.remove("active");
-    });
-  });
-});
-
-/* scroll to remove the sidebar */
-var prevScrollpos = window.pageYOffset;
-window.onscroll = function() {
-  var currentScrollPos = window.pageYOffset;
-  if (prevScrollpos < currentScrollPos) {
-    document.querySelector(".sidebar-menu").classList.remove("active");
-  } else {
-    document.querySelector(".sidebar-menu").classList.remove("active");
-  }
-  prevScrollpos = currentScrollPos;
-}
