--- old-assets/style.css
@@ -1,418 +1,0 @@
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-
-body {
-  background-image: linear-gradient(to left, rgb(10, 0, 73), rgb(0, 0, 0));
-  overflow-x: hidden;
-  color: rgb(255, 255, 255);
-  font-family: sans-serif;
-}
-
-.body::-webkit-scrollbar {
-  display: none;
-}
-
-.auto-typed {
-  background: -webkit-linear-gradient(#e4d6fb, #8400ff);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
-
-.section_title {
-  background: -webkit-linear-gradient(#e4d6fb, #8400ff);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  font-size: 4rem;
-  font-weight: bold;
-}
-
-#bg {
-  position: fixed;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  z-index: -1;
-}
-
-.input_container {
-  display: flex;
-  width: 100%;
-  justify-content: center;
-}
-
-.date_input {
-  width: 700px;
-}
-
-.apod_img_container {
-  width: 500px;
-  height: 500px;
-}
-
-.img {
-  width: 500px;
-  height: 400px;
-}
-
-.info {
-  width: 500px;
-  height: 100px;
-}
-
-#apod_img {
-  border-radius: 25px;
-}
-
-.apod_img{
-  position: relative;
-  height: 80%;
-}
-
-#apod_img:hover {
-  color: white;
-  box-shadow: 0 5px 15px #9000ff;
-}
-
-.hero_img {
-  border-radius: 50%;
-}
-
-.hero_img:hover {
-  color: rgb(191, 0, 255);
-  box-shadow: 0 5px 15px #9000ff;
-}
-
-.rover_container {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-}
-.rover_section_text {
-  width: 50%;
-}
-
-.mars_illus {
-  border-radius: 50%;
-  margin-right: 30px;
-}
-
-.mars_illus:hover {
-  color: rgb(191, 0, 255);
-  box-shadow: 0 5px 15px #ff0099;
-}
-
-.rover_display {
-  height: 70vh;
-  width: 100%;
-  display: none;
-  justify-content: center;
-}
-
-.inner_conatiner {
-  height: 100%;
-  width: 75%;
-  border-radius: 10px;
-  border: #e4d6fb;
-  border-style: solid;
-  border-width: 1px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.image_div {
-  width: 500px;
-  height: 500px;
-  margin-right: 50px;
-}
-
-.rover_popup_image_mobile {
-  width: 500px;
-  height: 500px;
-  border-radius: 25px;
-}
-
-#roverImg:hover {
-  color: rgb(191, 0, 255);
-  box-shadow: 0 5px 15px #9000ff;
-}
-
-.info_div {
-  width: 500px;
-  height: 500px;
-  background-color: aquamarine;
-}
-
-.info_container {
-  display: flex;
-}
-
-/* CONTACT US SECTION STYLING */
-
-.contact-form {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.contact-form .left {
-  flex: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.contact-form .right {
-  flex: 1;
-  padding-left: 20px;
-}
-
-.right .form-control {
-  width: 100%;
-  max-width: 600px; /* Adjust the maximum width as needed */
-}
-
-.contact_img {
-  border-radius: 50%;
-}
-
-.contact_img:hover {
-  color: white;
-  box-shadow: 0 5px 15px #9000ff;
-}
-
-/* Additional styling for responsiveness */
-@media (max-width: 768px) {
-  .contact-form {
-    flex-direction: column;
-  }
-
-  .contact-form .right {
-    padding-left: 0;
-    margin-top: 10px; /* Adjust as needed */
-  }
-}
-
-/* Info card in mars section */
-
-@import url("https://fonts.googleapis.com/css?family=Poppins:100,100i,200,200i,300,300i,400,400i,500,500i,600,600i,700,700i,800,800i,900,900i&subset=devanagari,latin-ext");
-
-.card {
-  border-radius: 10px;
-  filter: drop-shadow(0 5px 10px 0 #ffffff);
-  width: 230px;
-  height: 130px;
-  background-color: #ffffff;
-  padding: 20px;
-  margin-bottom: 5px;
-  margin-top: 5px;
-  margin-right: 10px;
-  position: relative;
-  z-index: 0;
-  overflow: hidden;
-  transition: 0.6s ease-in;
-}
-
-.info1 {
-  margin-right: 30px;
-}
-
-.card::before {
-  content: "";
-  position: absolute;
-  z-index: -1;
-  top: -15px;
-  right: -15px;
-  background: #7952b3;
-  height: 220px;
-  width: 25px;
-  border-radius: 32px;
-  transform: scale(1);
-  transform-origin: 50% 50%;
-  transition: transform 0.25s ease-out;
-}
-
-.card:hover::before {
-  transition-delay: 0.2s;
-
-  transform: scale(40);
-}
-
-.card:hover {
-  color: #ffffff;
-}
-
-.card p {
-  padding: 10px 0;
-}
-
-.img {
-  display: flex;
-  justify-content: center;
-}
-
-
-
-@media (max-width: 765px) {
-  /*apod section optimized for mobile devices*/
-
-  .section_title_mobile {
-    font-size: 2rem;
-  }
-
-  .img {
-    width: 100%;
-  }
-
-  .apod_img {
-    height: 350px;
-    width: 350px;
-  }
-
-  /*mars illus optimized for mobile devices*/
-
-  .rover_section_img {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding-bottom: 2rem;
-  }
-  .mars_illus {
-    width: 70%;
-  }
-
-  .rover_section_text {
-    width: 100%;
-  }
-
-  .rover_section_para {
-    text-align: center;
-    font-size: 1rem;
-    padding: 10px;
-  }
-
-  .rover_container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-  }
-
-  .rover_display {
-    display: none; /*change to implement this feature*/
-  }
-
-  .override-display {
-    width: 90%;
-  }
-
-  /* optimized hero_img for mobile devices*/
-  .hero_img {
-    width: 70%;
-  }
-
-  /*contact section optimized for mobile*/
-  .contact_img {
-    width: 70%;
-  }
-
-  /*rover popup display for mobile*/
-  .inner_conatiner {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    height: 30rem;
-    width: 98%;
-    border-radius: 10px;
-    border: #e4d6fb;
-    border-style: solid;
-    border-width: 1px;
-  }
-
-  .rover_popup_image_mobile {
-    height: 80%;
-    width: 80%;
-  }
-  .image_div {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 10rem;
-    width: 80%;
-  }
-
-  .card {
-    height: 5rem;
-    width: 20rem;
-  }
-
-  .info_container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .info-gen {
-    width: 20%;
-  }
-}
-
-.rover_container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.mars_illus {
-  height: 80%;
-}
-
-/* sidebar for mobile devices*/
-
-.sidebar-icon {
-  position: absolute;
-  top: 80px;
-  right: 20px;
-  z-index: 10;
-}
-
-.sidebar-list-item {
-  list-style-type: none;
-  padding-bottom: 8px;
-}
-
-@media (min-width: 990px) {
-  .sidebar-icon {
-    display: none;
-
-  }
-}
-
-.sidebar-contact {
-  display: flex;
-  justify-content: center;
-}
-.sidebar-menu {
-  position: fixed;
-  top: 80px;
-  right: -200px; 
-  width: 160px;
-  height: 160px;
-  background-color: inherit;
-  transition: right 0.3s ease;
-  z-index: 5; 
-}
-.sidebar-menu.active {
-  right: 0; 
-}
-
-.rem-default {
-  text-decoration: none;
-  color: inherit;
-  cursor: pointer;
-}
